check existence of config and pubkey

+++ dosenet.sh
@@ -11,8 +11,21 @@
 ### END INIT INFO
 HOME=/home/pi
 DOSENET=$HOME/dosenet-raspberrypi
+
 CONFIGFILE=$DOSENET/config.csv
+if [ ! -f $CONFIGFILE ]
+then
+  # no config file. exit with a user-defined exit code, 64
+  exit 64
+fi
+
 PUBLICKEY=$DOSENET/id_rsa_lbl.pub
+if [ ! -f $PUBLICKEY ]
+then
+  # no publickey. exit with a user-defined exit code, 65
+  exit 65
+fi
+
 LOG=/home/pi/sender.log
 echo "DoseNet script called @ " > $LOG
 date >> $LOG
